requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
boto3==1.24.46
botocore==1.27.46
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.5
Django==4.0.6
django-allauth==0.51.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.0
django-filter==22.1
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
jmespath==1.0.1
Markdown==3.4.1
oauthlib==3.2.0
Pillow==9.2.0
pip==22.1.2
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyotp==2.6.0
python-dateutil==2.8.2
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
sendgrid==6.9.7
setuptools==63.1.0
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
toml==0.10.2
twilio==7.12.0
urllib3==1.26.11
wheel==0.37.1
whitenoise==6.2.0
zipp==3.8.1
